Fix stack overflows, use *Node.id() instead of ref
All the tests in `test/run.sh` pass now.

Using procout I was able to debug the output of
derive::query::derive_query_db. What I saw was that key
comparison functions were "aliasing" in the caches, so it kept
retrieving the same scope over and over until the stack
overflowed. Looking at the code I saw PartialEq and Hash being
derived for keys with the arg signature of `&dyn *Node`, so any
of the different Node traits. While debugging I could see that
the references were identical for different `Node::id` values.
This makes some sense, Rust doesn't guarantee that a reference
to an object remains unchanged unless you `Pin` it. It looks
like it kept reusing the same memory in different iterations
of a loop. The solution I implemented replaces the derived
`PartialEq` and `Hash` for keys with a single arg that is a
`&dyn *Node`, and instead generates implementations that use the
`Node::id()` instead of the reference.

I'm not sure this is what is expected. I see comments in the code
about moving away from node ids. But unless you put `Box::pin`
objects into the cache you need to use a memory independent
id like `Node::id()`.